# rmn-fe/src/api/models/CreateMenuItemDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateMenuItemDTO = {
    categoryId?: number;
    itemName?: string | null;
    description?: string | null;
    basePrice?: number;
    thumbnail?: string | null;
};